Add an event listener.
We need to be able to listen to incoming events from Bukkit in order to
tell the rest of Mineground about them.

No tests, since this is just a simple layer of plumbing between Bukkit
and Mineground. There will be some logic, but the effort of mocking that
out really isn't worth it.