Add the CommandManager
Mineground features will have the ability to declare commands by
annotating a method in their class. The CommandManager provides for the
plumbing allowing that to happen.

CommandManager injects commands directly into Bukkit's own command
system, which feels awkwardly fragile to me. Unfortunately, there
doesn't seem to be a better way for defining commands at runtime, so
we'll stick with this for the time being. The backup plan would be to
define every single command in plugins.yml after all.