Poll for database results using a scheduled task
Since we'll be executing database queries on a separate thread, the main
thread will need to poll for finished results in order to settle the
promises.

Do this by creating a repeating synchronous task on Bukkit's scheduler,
to be ran every 2 server ticks. Bukkit servers run at ~20 ticks per
second, so the expected maximum delay in delivering database results is
about 100ms.