Initial version of the DatabaseResult class
This class will hold the results retrieved from the database, including
meta information such as the number of retrieved rows, number of
affected rows (for UPDATE queries) and the inserted Id for primary keys
in INSERT queries.

Documentation in the class and it holding the actual data has yet to
follow.