User-facing messages should be defined in mineground.yml
Add a Message class which can load them, and process macros in the
messages which should be replaced with actual values. There is a
terrible file encoding hack in there :-(. Start using it throughout the
plugin.

Also make the Color and TextDecoration classes non-instantiable. There
is no need for instances.