Generalize SimplePasswordHash, warp by world hashes
Generalize SimplePasswordHash to SimpleHash since it's not just used for
passwords anymore.

Also start saving warps by hashes of the world UUIDs rather than their
name, which will be more stable when creating a new world. While this
takes away one of the benefits of UUIDs (128-bit keys with little to no
repetition), a 32-bit key is much faster in the database, and we won't
create more than 4 billion worlds anyway. The UUID of a world is
genuinely random, and not based on the world's name.