Mineground should depend on Bukkit instead of CraftBukkit; fix JavaDoc errors
CraftBukkit are the implementation details, whereas Bukkit is the API
surface. In order to keep us from accidentally depending on
implementation details, let's depend on Bukkit for now.

Also fix a bunch of JavaDoc issues. I've significantly increased the
warning level in my Eclipse for the purpose of keeping stuff up to date
:-).