Enable commands to implement tab-completion
Introduce the @CommandCompletionHandler as a method-annotation usable in
Features (and feature components) to handle player auto-complete
requests. Methods having this annotation need to return a list of
strings, which can serve as suggestions.

This patch also implements it for the basic options for /warp. I haven't
yet figured out how to synchronously get the list of available warps...
:-).