CommandManager to understand remotes, expose /kick, /note and /notes
This patch updates the CommandManager to understand the concept of a
remote command executer.

The "kick", "note" and "notes" command will be exposed as the initial
remote commands. The IRC Manager will map these to "!kick", "!note" and
"!notes", using the same implementation as the in-game and console
version of the command.